requirements.txt
albumentations==1.3.1
entmax==1.1
huggingface_hub==0.19.4
munch==4.0.0
onnxruntime==1.16.3
opencv_python_headless==4.8.1.78
Pillow==10.1.0
PyYAML==6.0.1
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
einops==0.7.0
onnx==1.16.0